Chnaged the functions names from react to url
