--- functions/reactblock/requirements.txt
@@ -1,3 +1,0 @@
-crowdstrike-foundry-function==1.1.1
-crowdstrike-falconpy
-pytz
