--- ui/pages/reactblocking/.gitignore
@@ -1,3 +1,0 @@
-node_modules/
-.DS_Store
-.idea/
