--- ui/pages/reactblocking/src/styles.css
@@ -1,36 +1,0 @@
-/* app custom styles */
-/* styles.css */
-/* Custom scrollbar for URL preview */
-textarea::-webkit-scrollbar {
-  width: 8px;
-}
-
-textarea::-webkit-scrollbar-track {
-  background: #f1f1f1;
-  border-radius: 4px;
-}
-
-textarea::-webkit-scrollbar-thumb {
-  background: #cbd5e1;
-  border-radius: 4px;
-}
-
-textarea::-webkit-scrollbar-thumb:hover {
-  background: #94a3b8;
-}
-
-/* Animation for status messages */
-@keyframes slideIn {
-  from {
-    transform: translateY(-10px);
-    opacity: 0;
-  }
-  to {
-    transform: translateY(0);
-    opacity: 1;
-  }
-}
-
-.status-message {
-  animation: slideIn 0.2s ease-out;
-}
