+++ ui/pages/urlblocking/package.json
@@ -1,0 +1,40 @@
+{
+  "name": "foundry-js-react-blueprint",
+  "version": "1.0.0",
+  "license": "MIT",
+  "private": true,
+  "type": "module",
+  "directories": {},
+  "dependencies": {
+    "@crowdstrike/falcon-shoelace": "^0.4.0",
+    "@crowdstrike/foundry-js": "^0.17.1",
+    "@crowdstrike/tailwind-toucan-base": "^5.0.0",
+    "@shoelace-style/shoelace": "^2.20.1",
+    "d3": "^7.9.0",
+    "plotly.js": "^3.0.1",
+    "react": "19.1.1",
+    "react-dom": "19.1.0",
+    "react-plotly.js": "^2.6.0",
+    "react-router-dom": "7.6.1"
+  },
+  "devDependencies": {
+    "@babel/core": "7.27.3",
+    "@babel/preset-react": "7.27.1",
+    "@rollup/plugin-babel": "6.0.4",
+    "@rollup/plugin-commonjs": "28.0.6",
+    "@rollup/plugin-json": "6.1.0",
+    "@rollup/plugin-node-resolve": "16.0.1",
+    "@rollup/plugin-replace": "6.0.2",
+    "@web/rollup-plugin-html": "2.3.0",
+    "postcss": "^8.5.4",
+    "rollup": "4.41.1",
+    "rollup-plugin-postcss": "^4.0.2"
+  },
+  "scripts": {
+    "build": "rollup -c ./config/rollup.config.js",
+    "watch": "npm run build -- --watch"
+  },
+  "engines": {
+    "node": ">=18.12.0"
+  }
+}
