+++ ui/pages/urlblocking/src/index.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html lang="en" style="width: 100%; height: 100%">
+<head>
+  <title>URL Category Blocking</title>
+  <link rel="stylesheet" href="../node_modules/@crowdstrike/tailwind-toucan-base/index.css" />
+  <link rel="stylesheet" href="../node_modules/@crowdstrike/falcon-shoelace/dist/style.css" />
+  <link rel="stylesheet" href="./styles.css" />
+</head>
+<body class="bg-ground-floor">
+  <div id="app">
+    <div class="flex items-center justify-center min-h-screen">
+      <div class="text-center p-4">
+        <h1 class="text-xl font-bold text-indigo-600 mb-2">
+          Loading Application...
+        </h1>
+        <p class="text-gray-600">Please wait while we initialize</p>
+      </div>
+    </div>
+  </div>
+  <script src="./app.js" type="module"></script>
+</body>
+</html>
